Remove legacy server module
