+++ src/server/runtime.rs
@@ -1,0 +1,267 @@
+//! Runtime control for [`WireframeServer`].
+
+use std::{io, sync::Arc};
+
+use futures::Future;
+use tokio::{
+    net::TcpListener,
+    select, signal,
+    time::{Duration, sleep},
+};
+use tokio_util::{sync::CancellationToken, task::TaskTracker};
+
+use super::{
+    PreambleCallback, PreambleErrorCallback, WireframeServer, connection::spawn_connection_task,
+};
+use crate::{app::WireframeApp, preamble::Preamble};
+
+impl<F, T> WireframeServer<F, T>
+where
+    F: Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+    T: Preamble,
+{
+    /// Run the server until a shutdown signal is received.
+    ///
+    /// Spawns the configured number of worker tasks and awaits Ctrl+C for shutdown.
+    ///
+    /// # Errors
+    ///
+    /// Returns an [`io::Error`] if accepting a connection fails.
+    pub async fn run(self) -> io::Result<()> {
+        self.run_with_shutdown(async {
+            let _ = signal::ctrl_c().await;
+        })
+        .await
+    }
+
+    /// Run the server until the `shutdown` future resolves.
+    ///
+    /// # Errors
+    ///
+    /// Returns an [`io::Error`] if accepting a connection fails during runtime.
+    ///
+    /// # Panics
+    /// Panics if [`bind`](Self::bind) was not called beforehand.
+    pub async fn run_with_shutdown<S>(self, shutdown: S) -> io::Result<()>
+    where
+        S: Future<Output = ()> + Send,
+    {
+        let listener = self.listener.expect("`bind` must be called before `run`");
+        if let Some(tx) = self.ready_tx {
+            if tx.send(()).is_err() {
+                tracing::warn!("Failed to send readiness signal: receiver dropped");
+            }
+        }
+
+        let shutdown_token = CancellationToken::new();
+        let tracker = TaskTracker::new();
+
+        for _ in 0..self.workers {
+            let listener = Arc::clone(&listener);
+            let factory = self.factory.clone();
+            let on_success = self.on_preamble_success.clone();
+            let on_failure = self.on_preamble_failure.clone();
+            let token = shutdown_token.clone();
+            let t = tracker.clone();
+            tracker.spawn(worker_task(
+                listener, factory, on_success, on_failure, token, t,
+            ));
+        }
+
+        select! {
+            () = shutdown => shutdown_token.cancel(),
+            () = tracker.wait() => {},
+        }
+
+        tracker.close();
+        tracker.wait().await;
+        Ok(())
+    }
+}
+
+async fn worker_task<F, T>(
+    listener: Arc<TcpListener>,
+    factory: F,
+    on_success: Option<PreambleCallback<T>>,
+    on_failure: Option<PreambleErrorCallback>,
+    shutdown: CancellationToken,
+    tracker: TaskTracker,
+) where
+    F: Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+    T: Preamble,
+{
+    accept_loop(listener, factory, on_success, on_failure, shutdown, tracker).await;
+}
+
+pub(super) async fn accept_loop<F, T>(
+    listener: Arc<TcpListener>,
+    factory: F,
+    on_success: Option<PreambleCallback<T>>,
+    on_failure: Option<PreambleErrorCallback>,
+    shutdown: CancellationToken,
+    tracker: TaskTracker,
+) where
+    F: Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+    T: Preamble,
+{
+    let mut delay = Duration::from_millis(10);
+    loop {
+        select! {
+            biased;
+
+            () = shutdown.cancelled() => break,
+
+            res = listener.accept() => match res {
+                Ok((stream, _)) => {
+                    spawn_connection_task(
+                        stream,
+                        factory.clone(),
+                        on_success.clone(),
+                        on_failure.clone(),
+                        &tracker,
+                    );
+                    delay = Duration::from_millis(10);
+                }
+                Err(e) => {
+                    let local_addr = listener.local_addr().ok();
+                    tracing::warn!(error = ?e, ?local_addr, "accept error");
+                    sleep(delay).await;
+                    delay = (delay * 2).min(Duration::from_secs(1));
+                }
+            },
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::{
+        Arc,
+        atomic::{AtomicUsize, Ordering},
+    };
+
+    use rstest::rstest;
+    use tokio::{
+        sync::oneshot,
+        time::{Duration, timeout},
+    };
+
+    use super::*;
+    use crate::server::test_util::{bind_server, factory, free_port};
+
+    #[rstest]
+    #[should_panic(expected = "`bind` must be called before `run`")]
+    #[tokio::test]
+    async fn test_run_without_bind_panics(
+        factory: impl Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+    ) {
+        let server = WireframeServer::new(factory);
+        let _ = timeout(Duration::from_millis(100), server.run()).await;
+    }
+
+    #[rstest]
+    #[should_panic(expected = "`bind` must be called before `run`")]
+    #[tokio::test]
+    async fn test_run_with_shutdown_without_bind_panics(
+        factory: impl Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+    ) {
+        let server = WireframeServer::new(factory);
+        let shutdown_future = async { tokio::time::sleep(Duration::from_millis(10)).await };
+        let _ = timeout(
+            Duration::from_millis(100),
+            server.run_with_shutdown(shutdown_future),
+        )
+        .await;
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_run_with_immediate_shutdown(
+        factory: impl Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+        free_port: std::net::SocketAddr,
+    ) {
+        let server = bind_server(factory, free_port);
+        let shutdown_future = async { tokio::time::sleep(Duration::from_millis(10)).await };
+        let result = timeout(
+            Duration::from_millis(1000),
+            server.run_with_shutdown(shutdown_future),
+        )
+        .await;
+        assert!(result.is_ok());
+        assert!(result.unwrap().is_ok());
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_server_graceful_shutdown_with_ctrl_c_simulation(
+        factory: impl Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+        free_port: std::net::SocketAddr,
+    ) {
+        let server = bind_server(factory, free_port);
+        let (tx, rx) = oneshot::channel();
+        let handle = tokio::spawn(async move {
+            server
+                .run_with_shutdown(async {
+                    let _ = rx.await;
+                })
+                .await
+                .expect("server run failed");
+        });
+        let _ = tx.send(());
+        handle.await.expect("server join error");
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_multiple_worker_creation(
+        free_port: std::net::SocketAddr,
+    ) {
+        let call_count = Arc::new(AtomicUsize::new(0));
+        let clone = call_count.clone();
+        let factory = move || {
+            clone.fetch_add(1, Ordering::SeqCst);
+            WireframeApp::default()
+        };
+        let server = WireframeServer::new(factory)
+            .workers(3)
+            .bind(free_port)
+            .expect("Failed to bind");
+        let shutdown_future = async { tokio::time::sleep(Duration::from_millis(10)).await };
+        let result = timeout(
+            Duration::from_millis(1000),
+            server.run_with_shutdown(shutdown_future),
+        )
+        .await;
+        assert!(result.is_ok());
+        assert!(result.unwrap().is_ok());
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_accept_loop_shutdown_signal(
+        factory: impl Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+    ) {
+        let token = CancellationToken::new();
+        let tracker = TaskTracker::new();
+        let listener = Arc::new(
+            TcpListener::bind("127.0.0.1:0")
+                .await
+                .expect("failed to bind test listener"),
+        );
+
+        tracker.spawn(accept_loop::<_, ()>(
+            listener,
+            factory,
+            None,
+            None,
+            token.clone(),
+            tracker.clone(),
+        ));
+
+        token.cancel();
+        tracker.close();
+
+        let result = timeout(Duration::from_millis(100), tracker.wait()).await;
+        assert!(result.is_ok());
+    }
+}
