+++ src/server/test_util.rs
@@ -1,0 +1,46 @@
+//! Test helpers shared across server modules.
+
+use std::net::{Ipv4Addr, SocketAddr};
+
+use bincode::{Decode, Encode};
+use rstest::fixture;
+
+use crate::app::WireframeApp;
+
+use super::WireframeServer;
+
+#[derive(Debug, Clone, PartialEq, Encode, Decode)]
+pub struct TestPreamble {
+    pub id: u32,
+    pub message: String,
+}
+
+#[fixture]
+pub fn factory() -> impl Fn() -> WireframeApp + Send + Sync + Clone + 'static {
+    || WireframeApp::default()
+}
+
+#[fixture]
+pub fn free_port() -> SocketAddr {
+    let addr = SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 0);
+    let listener = std::net::TcpListener::bind(addr).expect("failed to bind free port listener");
+    listener
+        .local_addr()
+        .expect("failed to read free port listener address")
+}
+
+pub fn bind_server<F>(factory: F, addr: SocketAddr) -> WireframeServer<F>
+where
+    F: Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+{
+    WireframeServer::new(factory)
+        .bind(addr)
+        .expect("Failed to bind")
+}
+
+pub fn server_with_preamble<F>(factory: F) -> WireframeServer<F, TestPreamble>
+where
+    F: Fn() -> WireframeApp + Send + Sync + Clone + 'static,
+{
+    WireframeServer::new(factory).with_preamble::<TestPreamble>()
+}
